+++ lib/my_device.dart
@@ -1,4 +1,5 @@
 library my_device;
 
+export 'src/connectivity_configure.dart';
 export 'src/device.dart';
 export 'src/responsive_breakpoints.dart';